+++ cmake-scripts/boomerang-utils.cmake
@@ -79,6 +79,43 @@ endfunction()
 
 
 #
+# Usage: BOOMERANG_ADD_CODEGEN(NAME <name> SOURCES <source files> [ LIBRARIES <additional libs> ])
+#
+function(BOOMERANG_ADD_CODEGEN)
+    cmake_parse_arguments(CODEGEN "" "NAME" "SOURCES;LIBRARIES" ${ARGN})
+
+    option(BOOMERANG_BUILD_CODEGEN_${CODEGEN_NAME} "Build the ${CODEGEN_NAME} code generator." ON)
+
+    if (BOOMERANG_BUILD_CODEGEN_${CODEGEN_NAME})
+        set(target_name "boomerang-${CODEGEN_NAME}Codegen")
+        add_library(${target_name} SHARED ${CODEGEN_SOURCES})
+
+        set_target_properties(${target_name} PROPERTIES LIBRARY_OUTPUT_DIRECTORY "${BOOMERANG_OUTPUT_DIR}/lib/boomerang/plugins/codegen/")
+        set_target_properties(${target_name} PROPERTIES RUNTIME_OUTPUT_DIRECTORY "${BOOMERANG_OUTPUT_DIR}/lib/boomerang/plugins/codegen/")
+
+        if (MSVC)
+            # Visual Studio generates lib files for import in addition to dll files.
+            set_target_properties(${target_name} PROPERTIES ARCHIVE_OUTPUT_DIRECTORY_DEBUG "${BOOMERANG_OUTPUT_DIR}/lib/boomerang/plugins/codegen/")
+            set_target_properties(${target_name} PROPERTIES RUNTIME_OUTPUT_DIRECTORY_DEBUG "${BOOMERANG_OUTPUT_DIR}/lib/boomerang/plugins/codegen/")
+            set_target_properties(${target_name} PROPERTIES ARCHIVE_OUTPUT_DIRECTORY_RELEASE "${BOOMERANG_OUTPUT_DIR}/lib/boomerang/plugins/codegen/")
+            set_target_properties(${target_name} PROPERTIES RUNTIME_OUTPUT_DIRECTORY_RELEASE "${BOOMERANG_OUTPUT_DIR}/lib/boomerang/plugins/codegen/")
+            set_target_properties(${target_name} PROPERTIES ARCHIVE_OUTPUT_DIRECTORY_RELWITHDEBINFO "${BOOMERANG_OUTPUT_DIR}/lib/boomerang/plugins/codegen/")
+            set_target_properties(${target_name} PROPERTIES RUNTIME_OUTPUT_DIRECTORY_RELWITHDEBINFO "${BOOMERANG_OUTPUT_DIR}/lib/boomerang/plugins/codegen/")
+            set_target_properties(${target_name} PROPERTIES ARCHIVE_OUTPUT_DIRECTORY_MINSIZEREL "${BOOMERANG_OUTPUT_DIR}/lib/boomerang/plugins/codegen/")
+            set_target_properties(${target_name} PROPERTIES RUNTIME_OUTPUT_DIRECTORY_MINSIZEREL "${BOOMERANG_OUTPUT_DIR}/lib/boomerang/plugins/codegen/")
+        endif (MSVC)
+
+        target_link_libraries(${target_name} Qt5::Core boomerang ${CODEGEN_LIBRARIES})
+
+        install(TARGETS ${target_name}
+            LIBRARY DESTINATION lib/boomerang/plugins/codegen/
+            RUNTIME DESTINATION lib/boomerang/plugins/codegen/
+        )
+    endif (BOOMERANG_BUILD_CODEGEN_${CODEGEN_NAME})
+endfunction()
+
+
+#
 # Usage: BOOMERANG_ADD_FRONTEND(NAME <name> [ SOURCES <source files> ] [ LIBRARIES <additional libs> ])
 #
 # Note: There must be a <name>decoder.h/.cpp and a <name>frontend.h/.cpp present
@@ -98,20 +135,6 @@ function(BOOMERANG_ADD_FRONTEND)
 
     target_link_libraries(${FRONTEND_NAME} Qt5::Core ${FRONTEND_LIBRARIES})
 endfunction(BOOMERANG_ADD_FRONTEND)
-
-
-#
-# Usage: BOOMERANG_ADD_CODEGEN(NAME <name> SOURCES <source files> [ LIBRARIES <additional libs> ])
-#
-function(BOOMERANG_ADD_CODEGEN)
-    cmake_parse_arguments(CODEGEN "" "NAME" "SOURCES;LIBRARIES" ${ARGN})
-
-    add_library(boomerang-${CODEGEN_NAME}Codegen STATIC
-        ${CODEGEN_SOURCES}
-    )
-
-    target_link_libraries(boomerang-${CODEGEN_NAME}Codegen Qt5::Core ${CODEGEN_LIBRARIES})
-endfunction(BOOMERANG_ADD_CODEGEN)
 
 
 #