+++ Changelog.md
@@ -9,6 +9,7 @@ v0.5.0 (in development)
 - Fixed: Unsigned integral types mistaken for signed integral types when parsing signature or symbol information.
 - Fixed: Memory leaks in parsers.
 - Feature: The x86 decoder now recognizes a larger subset of the x86 instruction set.
+- Feature: Added support for Code Generator plugins.
 - Improved: Performance of decoding x86 instructions.
 - Improved: General processing of overlapped registers (not just hard-coded ones).
 - Improved: Better high level code output quality for x86 binaries due to more instructions being recognized.